requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.25
flask==3.0.1
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
flask-cors==4.0.0
